requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.27.2
structlog==24.4.0
celery[redis]==5.4.0
redis==5.2.1
aiogram==3.15.0
python-multipart==0.0.19
reportlab==4.2.5
openpyxl==3.1.5
boto3==1.35.84
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
email-validator==2.2.0